Remove the ability to set Upload and Download rate caps from footer
This code introduced a bug that caused a rate limit to be set arbitrarily
to 1 Gbps, overriding the one in the rtorrent config file. While it was
fairly easy to fix that bug, we have instead decided to remove this
feature.

In a server context, the default rates in the context menu make little
sense. I also suspect there's a larger number of people who would set
a rate limit using this feature accidentally than intentionally.

So from now on, client-wide settings should only be set in the
preferences pane. This puts this feature an extra click away for users
who need it but help us better cater to the majority use case.