Legacy compatibility removal (WIP)
* Specify a minimum xmlrpc-c version, and drop detection for bugs in very old versions of it
* Update a large number of xmlrpc calls to use the current name, removing the need for aliases
* Make alises unconditional with the goal of removing them eventually

We're going to pursue the opposite approach of what ruTorrent has done so far. They use the old
name in most places transparently converting it to the new name on newer rtorrent versions. We
will always use the newest name and convert it to the old one for older rtorrent versions
(identified by the API version, rather than the release version).