Remove UIEffects option
This setting never made much sense. There was no animation behavior
when resizing panes until we added it recently but there was a
pointless animation available for opening windows. It doesn't add
to the interface in a meaningful way, so it's getting the axe in
favour of faster, simpler behavior.